Final gate: cross-phase cleanup
encodePathSegments extracted (four inline copies); write-time mention
snippets share the feed's surrogate-safe truncateSnippet and constant;
FEED_* constants module-private; feedRowPath relocated to feedState
(dashboard's decision layer) restoring mentions.ts to its charter.
