+++ packages/web/src/lib/mentions.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, test } from 'bun:test'
-import { feedRowPath, type MentionUser, insertMention, mentionQuery, notificationHref } from './mentions'
+import { type MentionUser, insertMention, mentionQuery, notificationHref } from './mentions'
 
 const ada: MentionUser = { id: 'u-ada', name: 'Ada Lovelace', email: 'ada@example.com' }
 const noName: MentionUser = { id: 'u-x', name: null, email: 'x@example.com' }
@@ -85,42 +85,6 @@ describe('C21 — notificationHref: deep-link into the viewer review rail', () =
     )
     expect(notificationHref({ siteLabel: 'acme/doc', filePath: 'notes#draft.html', threadId: 't1' })).toBe(
       '/acme/doc/notes%23draft.html?thread=t1&review=1',
-    )
-  })
-})
-
-describe('C5.4 — feedRowPath: hide redundant root-file paths', () => {
-  test('hides index.html', () => {
-    expect(feedRowPath({ filePath: 'index.html', siteSlug: 'anything' })).toBeNull()
-  })
-
-  test('hides a lone file whose basename matches the site slug', () => {
-    expect(feedRowPath({ filePath: 'report.html', siteSlug: 'report' })).toBeNull()
-  })
-
-  test('slugifies a spaced basename before matching the site slug', () => {
-    expect(feedRowPath({ filePath: 'Q3 Report.html', siteSlug: 'q3-report' })).toBeNull()
-  })
-
-  test('shows a nested path even when its basename could match', () => {
-    expect(feedRowPath({ filePath: 'charts/revenue.html', siteSlug: 'revenue' })).toBe(
-      'charts/revenue.html',
-    )
-  })
-
-  test('matching is extension-agnostic', () => {
-    expect(feedRowPath({ filePath: 'report.md', siteSlug: 'report' })).toBeNull()
-  })
-
-  test('matching is case-insensitive through slugify', () => {
-    expect(feedRowPath({ filePath: 'REPORT.HTML', siteSlug: 'report' })).toBeNull()
-  })
-
-  test('a hidden display path is still carried by notificationHref', () => {
-    const item = { filePath: 'report.html', siteSlug: 'report' }
-    expect(feedRowPath(item)).toBeNull()
-    expect(notificationHref({ siteLabel: 'docs/report', filePath: item.filePath, threadId: 't1' })).toBe(
-      '/docs/report/report.html?thread=t1&review=1',
     )
   })
 })
